# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Peak Academy API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
...
  /api/study-rooms/join-random:
    post:
      summary: Join random study room
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Joined room
  /api/study-rooms/{roomId}/leave:
    post:
      summary: Leave study room
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Left room
...
  /api/reports/parent/{studentId}/pdf:
    get:
      summary: Generate parent student report PDF
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF generated
  /api/reports/parent/weekly-email-subscriptions:
    post:
      summary: Configure weekly report emails
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription saved
  /api/reports/parent/weekly-email-run:
    post:
      summary: Trigger weekly report email run
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Run completed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT